pyproject.toml
[project]
name = "gensearch"
version = "0.1.0"
description = ""
authors = [
    {name = "dixithak"}
]
packages = [
    {include = "gensearch"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "streamlit (>=1.51.0,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "langchain (>=1.0.7,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "faiss-cpu (>=1.13.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "openai (>=2.8.1,<3.0.0)",
    "boto3 (>=1.40.76,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-openai (>=1.0.3,<2.0.0)",
    "langchain-google-genai (>=3.1.0,<4.0.0)",
    "langchain-anthropic (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
